changes
I apologize for the confusion in my previous response. You're right; we should use a similar process as we did for the navigation menu to keep things consistent. Here's the correct way to translate the footer menu:

Update the en.yaml and es.yaml files with identifiers and translations for the footer menu items:
In en.yaml:

yaml
Copy code
#... (keep the existing content)

- id: press_media_footer
  translation: "Press & Media"

- id: about_us_footer
  translation: "About Us"

- id: get_involved_footer
  translation: "Get Involved"

- id: terms_conditions_footer
  translation: "Terms & Conditions"

- id: privacy_policy_footer
  translation: "Privacy Policy"
In es.yaml:

yaml
Copy code
#... (keep the existing content)

- id: press_media_footer
  translation: "Prensa y Medios"

- id: about_us_footer
  translation: "Sobre nosotros"

- id: get_involved_footer
  translation: "Participa"

- id: terms_conditions_footer
  translation: "Términos y condiciones"

- id: privacy_policy_footer
  translation: "Política de privacidad"
Update the config.toml file with the identifiers for the footer menu items:
toml
Copy code
#... (keep the existing content)

# footer menu
[[menu.footer]]
  identifier = "press_media_footer"
  url = "press/"
  weight = 1

[[menu.footer]]
  identifier = "about_us_footer"
  url = "about-us/"
  weight = 2

[[menu.footer]]
  identifier = "get_involved_footer"
  url = "get-involved/"
  weight = 3

[[menu.footer]]
  identifier = "terms_conditions_footer"
  url = "terms-conditions/"
  weight = 4

[[menu.footer]]
  identifier = "privacy_policy_footer"
  url = "privacy-policy/"
  weight = 5
Update the footer template (layouts/partials/footer.html) to use the translated menu items:
html
Copy code
<ul class="navbar-nav ml-auto">
  {{ range site.Menus.footer }}
  <li class="nav-item">
    <a class="nav-link" href="{{ .URL | relLangURL }}">{{ i18n .Identifier }}</a>
  </li>
  {{ end }}
</ul>
This should correctly translate the footer menu items on the Spanish site.

To update the names of any pages in the future, you can follow these steps:

Update the name in the content file's front matter (e.g., the title field in the corresponding Markdown file).

Update the name in the en.yaml and es.yaml files for the corresponding menu item identifier.

By updating these two places, you'll ensure that the page name is updated everywhere on the website.